Enable project reference to package dependency conversion for test bu…
* Enable project refrence to package dependency conversion in test builds

Opts some tests out of the process with "KeepAllProjectReferences"
Opts some project references out of conversion with "KeepProjectReference"
pulls in maririos change to sync.cmd
Respond to PR feedback

Note: Generated project.json's fail to restore because of missing test-runtime
property in frameworks which now have dependencies but previously did not.

* Move global.json to root so that generated project.json's can find it for the test-runtime

* System.Reflection.Context does not run on dnxcore50, but the project.json has that specified as the framework.
Can't convert this to netcore50 because there is no current test-runtime for netcore50.  Leave project reference
when doing a test build against packages so that it can use the netcore50 implementation directly.