Fix Process.StartTime on Unix on long-running systems
The current implementation of Process.StartTime on Unix uses Environment.TickCount to get a number of ticks since boot.  This value overflows after 24.8 days, resulting in incorrect StartTime values on systems that have run at least that long.

As at least a temporary fix, this commit switches to using Stopwatch, which uses the same underlying APIs as Environment.TickCount, but which uses long instead of int and thus doesn't suffer from the same limitation.  (To make this more robust, we should consider putting a GetBootTime or similar method into the native shim.)