Improve perf of Console.ResetColor()
Avoid the call to GetBufferInfo when we know we've already successfully called it.  ResetColor doesn't need any data from GetBufferInfo, it just uses it to know whether output is redirected, whether there is a console window, etc.  This a bit more than doubles the throughput of ResetColor. 95% of the remaining time is in SetConsoleTextAttribute, so there's no much more that can be done to meaningfully improve the throughput here.  And ForegroundColor/BackgroundColor can't do something similar, as they need data returned from GetBufferInfo.