Fix a potential overflow issue in Environment.TickCount tests.
Environment.TickCount can overflow after 25 days. The test now attempts to
take that into consideration by taking the difference between the current
TickCount value and the original. If the new TickCount value has
overflown, then the subtraction operation will also overflow, resulting in
a positive value. There is an implicit assumption here that the new TickCount
value will not overflow all the way to, for example, 0, because the values
are read in a loop with a one-second timeout. If such a thing occurred, we
would want to see an error.