Fix HttpClient package for NETCore50
Recent refactoring broke asset selection for HttpClient on NETCore50.

I had removed PackageTargetRuntime in order to use PackageDestination
instead but this caused the RID-specific implementation to be chosen for
NETCore50:
https://github.com/ericstj/corefx/commit/0a8c4fb115246f80bcc86ddb88797bfd1c7271f7#diff-d6c266b4814b075db300e51936fdf722L17

Instead we need to make sure that we package the NETCore50 build
as RID-specific as well so that it can take precedence of the RID-
specific NETStandard implementation.