Fail more gracefully if the user specifies a receive buffer for AcceptAsync
On Windows, AcceptAsync can optionally receive an initial buffer-full of data from the newly connected socket, via the magic of AcceptEx.  We currently have no emulation of this feature on Unix.  Instead, we fail an assert(in Debug builds, or report erroneous results in Release builds.

This change causes AcceptAsync to throw PlatformNotSupportedException if the user has supplied a buffer to receive data.  In the future, we can consider implementing this feature, but for now it's better to fail correctly rather than succeed erroneously.

I've also added a couple of new test cases (to check that this works on Windows, and does *not* work on Unix), and some new asserts to make sure this doesn't come up elsewhere.