Fix CurlResponseStream.Dispose to cancel the request
We weren't canceling the request when prematurely Dispose'ing the response stream.  Because of how asynchronous transfers work, where we need to pause the connection until a reader has requested data, if no reader ever requested data (e.g. the stream was disposed without reading anything) then the connection could remain paused indefinitely, and it would sit around until the HttpClient timeout kicked in and canceled it.  This could lead to unbounded growth of the number of sockets being used, as sockets couldn't be pooled or cleaned up due to still being associated with an active download.  The fix is simply to cancel the request in Dispose if the request isn't already completed.