Remove some redundances from expression interpreter
TryGetParameterOrReturnType and IndexIsNotReturnType are only called
from code for the FEATURE_FAST_CREATE switch. Have them only compile
for that switch.

LightCompiler.EmitLiftedNullCheck and LightDelegateCreator.HasClosure
are never used. Remove.

Result of call to TypeUtils.GetNullableType(node.Type) in
CompileBinaryExpression is never used and it has no side-effects. Remove.

IsNullableOrReferenceType is never used, but there are two places where
doing so would be a slight improvement. Use it.

ExceptionHelpers.prevStackTraces is only used if FEATURE_STACK_TRACES
is set. Have it only compile in that case.