Add Support for EnvelopedCms (Step 1: Empty Package.)
This class has been API reviewed and approved
   https://github.com/dotnet/corefx/issues/7830

For this first PR, the assembly has the actual code elided. As soon as this is merged, a followup PR will add the code.

For time/cost reasons, only Windows is supported for now. Unix platforms get a PlatformNotSupported shim. OpenSSL support is planned for later and the code is structured with a Pal structure similar to what we used for X509.

Platform targets are netcore50, net46, net461