Change InitializeStackTrace helper to store trace in Data
My previous attempt to get a better stack trace in exceptions was insufficient.  I was throwing/catching the exception to initialize the StackTrace, but I neglected the fact that the stack trace is filled in as the exception propagates, and so throwing/catching in the same frame means that the stack trace just stores that one frame, which isn't helpful.  Instead, this changes the helper to store the stack trace obtained via Environment.StackTrace into the Exception's Data dictionary.  This means the exception doesn't show up in a e.ToString() or e.StackTrace rendering, but it's at least stored for debugging purposes.