Add StrongToWeakReference<T> type to Common
CurlHandler needs to be able to maintain a reference to state that is strong at some points but weak at others, such that the referenced state can be kept alive while strong but allowed to be collected while weak.  This helper class allows for easy transitioning between the two states.