Support MemoryMappedFiles on UWP
The only thing preventing us from supporting this was switching to use
some UWP-specific win32 functions for memory mapped files.  This adds a
target for netcore50 to cross compile and use the *FromApp variants of
those win32 APIs.

I had to exclude a call to GlobalMemoryStatusEx because that was not
available, but the only purpose of that call was to get a better error
message.