Fix tarring on OSX
For some reason, there is automatic quoting of --exclude=*.Tests on OSX.  I am not sure whether this is coming from their shell implementation or Jenkins, but tar treats this as a directory sent to tar, which doesn't work.  Avoid doing this on OSX for now (though this increases the amount of data tar'd by about 3x).