Ensure Console's KeyFormatToConsoleKey dictionary is initialized
If the terminfo file for the current TERM can't be located, the TerminalFormatStrings cache currently doesn't initialize its string-to-ConsoleKeyInfo mapping dictionary.  Attempting to read from the console may then fail with a NullReferenceException when the code parsing the input line tries to access the dictionary.

The one-line fix is just to ensure the dictionary is always initialized.  I also added a few missing readonlys.