Add DefaultCredentials scenario tests
These tests require the client and server to be on a Windows Active Directory domain. These tests are Windows specific. Currently this cannot be done in the CI system. So, these tests are only enabled when an environment variable is set which points to the test server.

There is a separate todo item #4989 to provide more automated scripts for setting up the test server. For now, there is a readme file with some details.

These tests will help support fixes for #8176 and related bugs.