Reduce paths in Expression.Constant()
The untyped overload determines a type just to have that type tested
in ConstantExpression.Make() resulting in the same constructor call
either way. Just call that constructor directly.

This leaves caller to Make(). Since its type tests overlap with those
in Make(), merge the logic in the caller, and remove Make().