Remove redundant ToString() overrides in instructions.
In many cases a class derived from Instruction has a ToString()
that results in the same string as the base implementation.

While this would be a reasonable optimisation in a hot path, ToString()
is mostly used in debug assertions in instructions, with the remaining
cases also being debugging-related, so it's worth cutting it out as
bloat.