Add tests for Condition rejecting open generic, byref or pointer type
Contributes to #8081.

Just tests that it isn't accepted along with null constants (typed as
object). Currently it is possible to create a ConditionalExpression
that violates this rule, but it requires an other expression type to
do so, so when the rest of #8081 is dealt with then coerced null would
be the only remaining possibilty that wouldn't have already thrown in
creating a child expression.