Update CI to prepare for Universal Runner changes
The universal runner changes going into buildtools soon will impact how
we do our cross build testing.

Previously, the build would produce a complete test layout, which we
could then copy over to another machine and then use corerun to invoke
xunit.

With the universal runner changes, this will change such that the build
instead generates a RunTests script which copies all the assets the
tests consumes from packages into the execution directory and then
invokes the tests.

This implies we need to do the following changes:

- We need to start to archive both the bin folder and the packages
  folder and copy them over (since the universal runner RunTests.sh needs
  the packages folder in order to deploy test assets)

- We need to pass additional arguments like Outerloop and IgnoreForCI to
  build itself, as the build process uses these when generating the
  RunTests.sh file, since the script includes the xunit invocation
  command.

This impacts any run where we build the tests and then run them on
different machine (i.e. Windows Nano, certain Linux runs).

These are forward compatible changes to CI definitions so when we go to
make the updates, the builds happen as we expect and the CI system can
check that we are doing the right things.