Modify Path.Combine to not treat slashes as roots
Path.Combine on all systems treats directory separators as the root. This is somewhat unexpected behavior on Windows and unhelpful on Unix. This commit changes the behaviors of all systems for path.Combine with path segments that begin with a DirectorySeparator and adds tests for the new behavior.

PR feedback