Override CleanAllProjects instead of Clean in build.proj
Prior to this change, build.proj overrode the Clean target
directly which did not honor TraversalCleanDependsOn.  This
meant other repo's could not customize the behavior of
a clean, either via Clean.cmd or 'build /t:clean'.

With this change, other repo's can add custom build targets
to TraversalCleanDependsOn to do pre- or post- clean handling.
WCF in particular requires this.