Fix Issue 2583: PfxTests.TestPrivateKey failure
Rebuilt PfxData with "openssl pkcs12" to remove
the bag attribute containing a preferred container name.
Now, crypto is forced to guidgen a container name even
on the first open so the odds of two cert instances
thinking they own the same key container are effectively
zero.

Opportunistically did the same to the other PFX's, provided
that openssl roundtripped the pfx with no issues.
But PfxData is the only one that affects this particular
test.