Add Stream.Begin/EndRead and Stream.Begin/EndWrite
While we want developers to implement and consume the ReadAsync/WriteAsync methods on Stream, not having the older Begin/End* methods causes portability problems for libraries (it also makes porting harder).

This commit adds back the virtual methods in the contract and provides an implementation for Stream in corert (coreclr already has implementations in mscorlib).  It also provides overrides on the streams defined in System.IO.dll.

Once a new package is published with this updated contract, I'll add overrides to the rest of our stream implementations in corefx as well as tests.