Clean up Globalization test for disabled Linux cases
There is differences between Windows and Linux/ICU in the colation area and it
is expected we see differnces when doing some string collation operations like
Compare, IndexOf, LastIndexOf, IsPrefix and IsSuffix.
The change here is to clean the test to have it run without any failure on all
platforms.