PLINQ - Fewer/Faster tests (#8451)
Peome earlier discussion, I've cut down on the number and running time of the PLINQ tests.

The majority of the cases are still/should be covered, the reductions have generally been in sizes and source types for unordered tests.  Overall, OuterLoop tests on my machine are running about 40% faster. The number of tests is down, but the combinatorial tests still represent the majority of the count.

There are also some minor fixes here and there - dealing with an instance of integer overflow in a helper method, for example.  some earlier discussion, I've cut down on the number and running time of the PLINQ tests.

The majority of the cases are still/should be covered, the reductions have generally been in sizes and source types for unordered tests.  Overall, OuterLoop tests on my machine are running about 40% faster. The number of tests is down, but the combinatorial tests still represent the majority of the count.  There are also some minor fixes here and there - dealing with an instance of integer overflow in a helper method, for example.