Include ECDsa in the netcore50 S.S.C.Algorithms implementation
When making the contract changes to add import/export for ECDsa the
netcore50 build got split off to a separate assembly.  This assembly was
supposed to match the 4.1 contract (which has ECDsa), but when excluding the
new EC* types it also lost ECDsa.

Since having more types in the implementation than required by the contract
isn't bad, just go ahead and include all the files in the netcore50 build.  This
brings ECDsa back, and restores the ability to use ECDsa (directly or indirectly)
in UWP applications.