Deterministically dispose CngKey.Handle values.
Fixes: https://github.com/dotnet/corefx/issues/6010

Dispose all the handles we're pulling out of CngKey (which
means the similarly named helper in Algorithms has to
duplicate the safehandle too since they feed the same
set of source-linked files. So we have to add a light-weight
DuplicateSafeNcryptKeyHandle to go with our light-weight
SafeNcryptKeyHandle.

Latching the key handle independently of the key is fragile
and tricky, especially with the key management split between
RSA, RSACng and CngAlgorithmCore. It also opens us
up to the "You forgot that the GC can claim objects
and their fields while one of their instance methods are executing"
bugs. (In fact, we probably had some of those bugs lurking
now that this commit fixes.) Not worth it given that we're doing
expensive crypto algorithms in between.

- (PR Note: added SetHandle(IntPtr.Zero) to ReleaseHandle
  override and squashed.)