Remove the calls to CryptoNative_GetStreamSizes
The StreamSizes concept is really around doing manual message framing with
Windows SChannel. Since OpenSSL libssl does all of the framing within the
library it's entirely opaque to us.

So, populate the structure with fixed values, and eliminate the call into the shim.

If there's a need for something fancier in the future, it'll require a different
function (which takes the SSL* as input to make the relevant fancy decisions).