Updating contracts for new UTF8 string marshalling APIs
We're adding a new API version which needs to be OOB on desktop.
Ensure we continue to build the 4.6.2 version.
Right now we're lying and saying .NETCoreApp1.0 supports the latest API.
Once we have a netcoreapp1.1/netstandard1.7 we'll need to update to use
that and either redistribute or cross compile for netstandard1.5/
netcoreapp1.0.