Create lookup table for IsDefault and TypeId
System.ComponentModel.TypeDescriptor expects certain attributes(defined
in System.ComponentModel.Primitives) to override IsDesfaultAttribute and
TypeId from System.Attribute. This is not available on .NET Core, so an
interface had been supplied for it instead. This design required the
projects to depend on each other rather than the ref assemblies, so this
removes that dependency by creating a lookup table that should cover
the majority of the cases.