Change FileStream async default back to false
FileStream's ctors allow the creator of the stream to specify whether asynchronous I/O should be used.  The default in the full framework is false.  When FileStream was ported to corefx, the default was changed to true.

Such a default of true can be beneficial, as it means that ReadAsync/WriteAsync calls on streams created without overriding that default will naturally use overlapped I/O on Windows, which can lead to better scalability and performance.  However, there are other ramifications, in particular for existing code and code written to be portable across .NET implementations.  For example:
- Using Read/Write on a stream created with useAsync=true ends up blocking as calls around Read/WriteAsync, making them more expensive than they'd otherwise be.  So existing code that uses a FileStream for simple read/writes may actually get slower.
- The SafeFileHandle exposed from the FileStream naturally is associated with that particular async setting, such that attempting to use it to construct a FileStream with a conflicting setting will fail with an exception.
- FileStreams constructed with useAsync=true have special behaviors around how the position of the stream is updated after an async operation is initiated, which means code written expecting one default may fail when run with a different default.
- Etc.
While there are benefits to changing the default, the cons outweigh the pros.  A developer can always opt-in to useAsync=true, but it's better to keep the default consistent across all implementations, especially for code written to be portable across those implementations.  This commit changes it back to be false.