Update S.S.C.OpenSsl contract
This elevates the minimum supported framework for S.S.C.OpenSsl to
netstandard1.6.

Let GenApi rewrite the contract file to include all of the new overrides, plus the new constructor.  Since the type is sealed the overrides aren't all that important, but the new constructor is only callable with the new contract.