Avoid using tmp subdirectories in Pipes and MMF on Unix
We we originally implemented the System.IO.Pipes and System.IO.MemoryMappedFiles libraries on Unix, they would end up persisting temporary files for potentially long durations, and in a manner that would make it likely for these files to persist even after a clean process shutdown.  As a result, we put these temporary files in subdirectories of TMPDIR, e.g. /tmp/.dotnet/corefx/maps/.

Since then, we've changed these implementations such that the files are much less persistent.  Files created by the MMF library exist only for a moment, created and then immediately unlinked.  And files created by the pipes library typically also exist for a very short period of time, between the time when the server starts listening and the client connects, after which the file is deleted.  This makes the subdirectories much less useful.

Further, the subdirectories cause permissions problems.  When mkdir is used to create a directory, the process' umask is automatically applied, such that even if we ask for 777 permissions, we're likely to actually create the directory as 775 (if we're a normal user) or as 755 (if we're root).  In either case, this causes problems when one user creates the temporary directory that another user then tries to open and write into.  There are workarounds for this, such as creating a temporary directory with mkdir, chmod'ing it to get the actual desired permissions, and then renaming that to the actual target directory desired, but that's complicated and leads to more potential points of unnecessary failure.

As a result of all of this, we can simply do away with these subdirectories, simplifying the code and leading to fewer issues in production.