Workaround missing InteropServices.PInvoke
System.Runtime.InteropServices.PInvoke was removed.  The problem with
this is that the previous build was still consuming reference assemblies
that typeforwarded members from InteropServices to
InteropServices.PInvoke.

This was only being hit for one assembly in the tree:
System.Net.Security's unix implementation.

This caused a dependency to be declared from the System.Net.Security
package that would never be satisifed.

We'll workaround this for this update, by downgrading the
System.Runtime.InteropServices.PInvoke contract.
This will work at runtime since all the types used happen to be
forwarded to mscorlib in both IS.PInvoke and IS.  I've also marked
this downgraded reference with exclude: compile.
Though not strictly needed here since test-runtime is already marked
the same, I wanted to be doubly sure that the downgraded reference
would never hit the compiler since it's compile assets would have type
conflicts with InteropServices.