Move tests that don't specify a target group their own directory.
Today when a test that specifies a target group is built it will place the
output for that target group in a subdirectory of the build with an empty
target group; this leads to a race condition in the build when
ArchiveTests is set to true. Instead, place each test build in its own
directory; and in cases where TargetGroup is empty use netstandard as the
subdirectory name.
Update the test runner scripts with the new default directory.
Pick up latest build tools for some fixes for the official build.