Implement Console support for netcore50
The Windows implementation currently used for netcore50 will fail the WACK check as it calls a number of APIs unsupported by UWP. This commit adds a separate netcore50 build that stubs out the API calls and replaces them with PlatformNotSupportedExceptions or no-ops where appropriate. The result is a platform-agnostic version of System.Console that can be used anywhere but is severely stripped of functionality.

- Where possible the behavior of a redirected Console operation will attempt to mimic that of a Unix redirected Console operation. Otherwise it will throw.
- Encoding is fixed to UTF8. Attempting to set the Console encoding will no-op.
- Added a section in the packaging build for the new netcore50 specific binary
- Input/Output redirection and subsequent reads/writes work as expected.