Update the System.AppContext library to call out that it supports 4.6.3
The entire surface area for AppContext will only be supported by Desktop in 4.6.3.

The 2 APIs that are currently baselined were introduced in 4.6.3.
But we don't have a targeting pack for that and we are using the 4.6.2. one which does not expose those APIs.

Once we have the targeting pack for 4.6.3. we are going to be able to remove the baseline file.