Delay HttpResponseMessage publishing until all request data sent
While it's not common, there are cases where an HTTP server may send response headers and body before all of the request content has been sent or received.  In such cases, we want CurlHandler to not publish the HttpResponseMessage until all of that request data has been sent, such that on both Windows and Unix the response message isn't published until all request headers and body has been sent and response headers have been received.  While this behavior may be relaxed in the future, it's critical to do this now, as the request content is Dispose'd of by HttpClient once the Task returned from SendAsync has completed.  If the SendAsync Task is completed before all of the request content has been sent, failures can easily occur due to Dispose'ing of the request content object while it's still being used.