Switch test-runtime to NETStandard.Library
Microsoft.NETCore.Console soon won't be built with the same versioning
as the rest of CoreFx: https://github.com/dotnet/core-setup/pull/14

We don't want to have to "know" which version of this package is
consistent with the CoreFx version we're picking up.

In addition to this Microsoft.NETCore.Console was too large: we really
only need the test-runtime to bring in a closure of dependencies for the
test runner and libraries.

In addition to this I've split up the validation rules for packages that
will soon start coming from other repos.

Add missing impl dependencies to tests

These dependencies are exclusive to the product assembly brought in via
project-reference.  Because our current restore process doesn't catch
transitive package dependencies these need to be added manually.

Fix up some of more project dependencies