Fix Environment.StackTrace throwing BadImageFormatException.
Issue #8866.

The main reason this was happening has been fixed on Windows on the coreclr side, but there is still a scenario Linux/OSX where a dev builds an app on Windows and copies it to Linux with an old pdb.  This will cause the same bad image format exception.

This fix wraps created the portable PDB reader in a try/catch (BadImageFormatException).

.