Reduce CredentialKey allocations
The AuthenticationType.ToUpperInvariant().GetHashCode() allocation
can be avoided in GetHashCode by using
StringComparer.OrdinalIgnoreCase.GetHashCode(AuthenticationType).

There's no need to include the length in the hash code, as the
hash code for the Uri is sufficient.

There's also no need to cache the hash code, which allows us to reduce the
size of CredentialKey.