Stop cross building other Linux distros
We should now be able to build natively on all Linux distros. Let's cut
out building some components on Windows and others on Linux and using
run-test.sh in favor of just doing everything on the target machine.

We also move to using CLR from packages instead of a live CLR.

I have also removed the outerloop runs we were doing with a checked
CoreCLR, it turns out that CoreCLR had similar runs and the failures are
more actionable to them.

Fixes #8860