Make UWP clrcompression package aot specific
This commit modifies the packaging of the ClrCompression native windows binary to work for all RIDs and frameworks. Previously, using the win10 package on a non-UWP program would fail becase there was only one native asset in the package: the App-container one. Becase of NuGet limitations we couldn't put regular binaries alongside the AppX one, so we distinguished AppX vs non-AppX by the win10 vs win7 RID. With this commit we'll only ship the AOT package with the APPX clrcompression that is built internally; all other RIDS and Frameworks will get the same clrcompression version that is built from the Open.