Ignore standard error when querying openssl
On some platforms, we get error text from openssl when we query for null
ciphers. The error message is written to standard error and printed in a
way such that MSBuild thinks xUnit has encounted an error and fails the
entire test suite.

To work around this issue, we'll just ignore standard error when we do
this query.

Fixes #9272